+++ piecrust/CHANGELOG.md
@@ -9,13 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `spent` field to `CallTreeElem` [#206]
+- Add `call_tree` to `CallReceipt` [#206]
+- Expose `CallTree` and `CallTreeElem` in the public API [#206]
+- Add `CallTreeIter` to improve iteration over call tree [#206]
 - Add `panic` import implementation [#271]
 - Add `Error::ContractPanic` variant [#271]
 
 ### Changed
 
+- Rename `StackElement` to `CallTreeElem` [#206]
 - Allow for multiple initializations on a new memory [#271]
 - Downcast `Error::RuntimeError` on each call boundary [#271]
+
+### Removed
+
+- Remove `CallStack` in favor of `CallTree` [#206]
 
 ## [0.10.0] - 2023-09-13
 
@@ -250,6 +259,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#216]: https://github.com/dusk-network/piecrust/issues/216
 [#213]: https://github.com/dusk-network/piecrust/issues/213
 [#207]: https://github.com/dusk-network/piecrust/issues/207
+[#206]: https://github.com/dusk-network/piecrust/issues/206
 [#202]: https://github.com/dusk-network/piecrust/issues/202
 [#201]: https://github.com/dusk-network/piecrust/issues/201
 [#181]: https://github.com/dusk-network/piecrust/issues/181